chore: rename plugins module to adapters for clarity
